requirements.txt
# 数据获取
akshare>=1.11.0
pandas>=1.5.0
numpy>=1.24.0

# 数据存储
sqlite3
redis>=4.5.0

# 技术分析
ta-lib>=0.4.25
talib>=0.4.21

# 数据可视化
matplotlib>=3.6.0
plotly>=5.17.0
seaborn>=0.12.0

# Web框架
flask>=2.3.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# 异步处理
celery>=5.3.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 数据验证
pydantic>=2.0.0

# 时间处理
pytz>=2023.3

# 机器学习（可选）
scikit-learn>=1.3.0
lightgbm>=4.0.0

# 测试
pytest>=7.4.0
pytest-mock>=3.11.0